Initial commit; project skeleton
This is the initial commit of the project, which is to provide a
portable library for micro-controllers that allows an LED and two
digital pins to be used for two way communications and sensing of the
ambient light. This is unlike my other Arduino project available at
<https://github.com/howerj/arduino> in that it is only a test bed for
the LED communications and not for everything Arduino related.